pyproject.toml
[project]
name = "trump-corpus"
version = "0.1.0"
description = "Unified corpus of Donald Trump's social media posts and public speeches for text mining."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "python-dateutil>=2.9",
    "tqdm>=4.66",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "bertopic>=0.16",
    "sentence-transformers>=3.0",
    "plotly>=5.22",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
]

[project.optional-dependencies]
dev = ["ruff>=0.5", "pytest>=8.0"]

[project.scripts]
trump-build = "trump_corpus.cli:build"
trump-refresh = "trump_corpus.cli:refresh"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trump_corpus"]

[tool.ruff]
line-length = 100
target-version = "py311"
